+++ services/pane/index.test.js
@@ -3,7 +3,6 @@ var dirname = __dirname.split('/').pop(),
   lib = require('./'),
   dom = require('@nymag/dom'),
   tpl = require('../tpl'),
-  ds = require('dollar-slice'),
   fixture = require('../../test/fixtures/tpl');
 
 describe(dirname, function () {
@@ -19,8 +18,6 @@ describe(dirname, function () {
       sandbox.stub(tpl, 'get');
       fixture.stubAll([
         '.kiln-pane-template',
-        '.preview-actions-template',
-        '.share-actions-template',
         '.filtered-input-template',
         '.filtered-items-template',
         '.filtered-item-template',
@@ -86,32 +83,6 @@ describe(dirname, function () {
         lib.close();
         fn([{header: header, content: innerEl, disabled: true}]);
         expect(document.querySelector('#pane-tab-1').classList.contains('disabled')).to.equal(true);
-      });
-    });
-
-    describe('openAddComponent', function () {
-      var fn = lib[this.title],
-        sandbox;
-
-      beforeEach(function () {
-        sandbox = sinon.sandbox.create();
-        sandbox.stub(ds);
-      });
-
-      afterEach(function () {
-        sandbox.restore();
-      });
-
-      it('opens a an add component pane', function () {
-        var options = {
-          field: { ref: null, path: null}, // parent data, passed to addComponent (we don't care about it here)
-          pane: document.createElement('div') // pane element, passed to addComponent (we don't care about it here)
-        };
-
-        lib.close();
-        fn(['foo'], options);
-        expect(document.querySelector('#pane-tab-1').innerHTML).to.equal('Add Component');
-        expect(document.querySelectorAll('.pane-inner li.filtered-item').length).to.equal(1);
       });
     });
   });
